pyproject.toml
[tool.poetry]
name = "pyconit"
version = "0.1.0"
description = ""
authors = ["sagie-shamay <sagie@twingate.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
webauthn = "^1.8.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
